phase-2: single source of truth for serving config; fix Epic encoder mismatch + float32 crash
Add apps/backend/platform_config.py as the only place per-platform serving constants live; app.py builds all predictors from it and derives /api/health from it. Epic now loads the matched *_epic encoder/stats (D-005). Cast model day-counts to float before timedelta to fix the XGBoost-path crash this surfaced (P9). Remove dead/divergent constants from train_models.py. Verified locally via Flask test client. (D-004, D-005, D-010)

+++ apps/backend/services/predictor.py
@@ -387,7 +387,7 @@ class GameServicePredictor:
                 "recently_appeared": recently_appeared,
                 "months_since_last": float(months_since),
                 "prediction_basis": "wait_time",
-                "projected_arrival": (datetime.now() + timedelta(days=predicted_months * 30)).strftime("%B %Y"),
+                "projected_arrival": (datetime.now() + timedelta(days=float(predicted_months * 30))).strftime("%B %Y"),
             }
         except Exception as e:
             print(f"Error in predict_repeat: {e}")
@@ -503,7 +503,7 @@ class GameServicePredictor:
             "publisher_avg_wait_days": float(pub_stats["pub_avg_days"]),
             "predicted_total_days": float(predicted_days_total),
             "metacritic_score_used": float(meta_score),
-            "projected_arrival": (datetime.now() + timedelta(days=days_remaining)).strftime("%B %Y"),
+            "projected_arrival": (datetime.now() + timedelta(days=float(days_remaining))).strftime("%B %Y"),
         }
 
     def predict(
